+++ components/reading/reading-type.tsx
@@ -4,7 +4,7 @@ import { Button } from "../ui/button"
 import { Card } from "../ui/card"
 import { Badge } from "../ui/badge"
 import { useTarot } from "@/contexts/tarot-context"
-import { ReadingConfig } from "../../app/[locale]/reading/page"
+import { ReadingConfig } from "../../app/[locale]/tarot/page"
 import { isFollowUpQuestion, getCleanQuestionText } from "@/lib/question-utils"
 import { useEffect, useState } from "react"
 import { useTranslations } from "next-intl"
